修改zk-Latency avg为小数时的数值转换异常问题

+++ km-common/src/main/java/com/xiaojukeji/know/streaming/km/common/bean/entity/zookeeper/fourletterword/parser/ServerCmdDataParser.java
@@ -54,7 +54,7 @@ public class ServerCmdDataParser implements FourLetterWordDataParser<ServerCmdDa
                     case "Latency min/avg/max":
                         String[] data = elem.getValue().split("/");
                         serverCmdData.setZkMinLatency(Long.valueOf(data[0]));
-                        serverCmdData.setZkAvgLatency(Long.valueOf(data[1]));
+                        serverCmdData.setZkAvgLatency(Float.valueOf(data[1]));
                         serverCmdData.setZkMaxLatency(Long.valueOf(data[2]));
                         break;
                     case "Received":
